# Making the tests faster.