--- a/tools/docker/cpu_tests.Dockerfile
+++ b/tools/docker/cpu_tests.Dockerfile
@@ -1,10 +1,12 @@
-FROM gcr.io/tensorflow-testing/nosla-ubuntu16.04-manylinux2010
+FROM python:3.5
 
-COPY build_deps/build-requirements.txt ./
-RUN pip3 install -r build-requirements.txt
+RUN pip install tensorflow-cpu==2.1.0
+
+RUN git clone https://github.com/abhinavsingh/setup-bazel.git
+RUN bash ./setup-bazel/setup-bazel.sh 1.1.0
 
 COPY requirements.txt ./
-RUN pip3 install -r requirements.txt
+RUN pip install -r requirements.txt
 
 COPY ./ /addons
 WORKDIR addons