@seanpmorgan 2020-02-10
> Keep in mind that the total time vary a lot in the CI, so it might not be fair to compare using that. This is why I compare the speed of the steps:
> 
> Step	before (s)	after (s)
> Downloading docker image	74	22
> installing tf	93	54
> installing bazel	0	8
> The speed of compiling and testing should in theory be the same.
> 
> I also think it's nice to start from scratch to build and run because:
> 
> * it helps other to understand how to build tensorflow addons from source locally. We don't start from a big docker image with who-knows-what already installed in it.
> * Here we have full control over the python version and bazel version for example.
> 
> We might want to do the same for the gpu and releases.

So for the release we need that specific docker image since it has devtoolset-7 and will build manylinux2010 compatible wheels. For testing I believe this is fine, but will need to review when I have a little time.