@gabrieldemarmiesse 2020-02-11
> To expand on this, we could (and it would be nice) to create our own image. But there are some things we have to get correct.

Sure thing. Building it is a first step :)